@@ -11,13 +11,17 @@
             "email": "laurent@bassin.info"
         },
         {
+            "name": "Trevor Rowbotham",
+            "email": "trevor.rowbotham@pm.me"
+        },
+        {
             "name": "Symfony Community",
             "homepage": "https://symfony.com/contributors"
         }
     ],
     "require": {
-        "php": ">=5.3.3",
-        "symfony/polyfill-mbstring": "^1.3",
+        "php": ">=7.1",
+        "symfony/polyfill-intl-normalizer": "^1.10",
         "symfony/polyfill-php72": "^1.10"
     },
     "autoload": {
@@ -30,7 +34,11 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "1.17-dev"
+            "dev-main": "1.22-dev"
+        },
+        "thanks": {
+            "name": "symfony/polyfill",
+            "url": "https://github.com/symfony/polyfill"
         }
     }
 }
