@@ -6,13 +6,13 @@
  * @group API
  * @since 0.1
  */
-class API_Output_Tests extends PHPUnit_Framework_TestCase {
+class API_Output_Tests extends PHPUnit\Framework\TestCase {
 
     protected $header_content;
 
     protected $header_status;
 
-    protected function tearDown() {
+    protected function tearDown(): void {
         yourls_remove_all_actions( 'content_type_header' );
         yourls_remove_all_actions( 'status_header' );
     }
@@ -24,7 +24,7 @@ class API_Output_Tests extends PHPUnit_Framework_TestCase {
      */
     public function test_output_return_or_echo() {
         $content = array( 'simple' => 'test' );
-        
+
         // return
         $this->assertSame( 'test', yourls_api_output( 'simple', $content, false, false ) );
         // echo
@@ -62,10 +62,10 @@ class API_Output_Tests extends PHPUnit_Framework_TestCase {
     public function test_output_content_headers( $type, $expected_header ) {
         yourls_add_action( 'content_type_header', array( $this, 'get_content_header' ) );
         yourls_add_action( 'status_header', array( $this, 'get_status_header' ) );
-        
+
         $content = array( 'hello' => 'test' );
         $result  = yourls_api_output( $type, $content, true, false );
-        
+
         $this->assertSame( $expected_header, $this->header_content );
     }
 
@@ -75,11 +75,11 @@ class API_Output_Tests extends PHPUnit_Framework_TestCase {
     public function status_code() {
         $success_code = mt_rand( 1, 10 );
         $error_code   = mt_rand( 1, 10 );
-        
+
         $content_success = array( 'statusCode'    => $success_code );
         $content_error   = array( 'errorCode'     => $error_code );
         $content_random  = array( 'thereIsNoCode' => rand_str() );
-        
+
         return array(
             array( $content_success, $success_code ),
             array( $content_error,   $error_code ),
@@ -95,9 +95,9 @@ class API_Output_Tests extends PHPUnit_Framework_TestCase {
      */
     public function test_output_status_headers( $content, $expected_status ) {
         yourls_add_action( 'status_header', array( $this, 'get_status_header' ) );
-        
+
         $result  = yourls_api_output( 'simple', $content, true, false );
-        
+
         $this->assertSame( $expected_status, $this->header_status );
     }
 
