@@ -6,15 +6,15 @@
  * @group signatures
  * @since 0.1
  */
-class Auth_Sig_Tests extends PHPUnit_Framework_TestCase {
+class Auth_Sig_Tests extends PHPUnit\Framework\TestCase {
 
     protected $backup_request;
 
-    protected function setUp() {
+    protected function setUp(): void {
         $this->backup_request = $_REQUEST;
     }
 
-    protected function tearDown() {
+    protected function tearDown(): void {
         $_REQUEST = $this->backup_request;
     }
 
