@@ -5,14 +5,14 @@
  *
  * @group db
  */
-class DB_Set_Tests extends PHPUnit_Framework_TestCase {
+class DB_Set_Tests extends PHPUnit\Framework\TestCase {
 
     protected $ydb_copy = null;
 
     /**
      * Make a copy of $ydb
      */
-    public function setUp() {
+    public function setUp(): void {
         $this->ydb_copy = yourls_get_db();
         yourls_set_db(null);
     }
@@ -20,7 +20,7 @@ class DB_Set_Tests extends PHPUnit_Framework_TestCase {
     /**
      * Restore original $ydb
      */
-    public function tearDown() {
+    public function tearDown(): void {
         yourls_set_db($this->ydb_copy);
     }
 
