@@ -5,7 +5,7 @@
  *
  * @group install
  */
-class Install_Tests extends PHPUnit_Framework_TestCase {
+class Install_Tests extends PHPUnit\Framework\TestCase {
 
     /**
 	 * Check if YOURLS is declared installed
@@ -79,20 +79,22 @@ class Install_Tests extends PHPUnit_Framework_TestCase {
 
 	/**
 	 * Test incorrect config provided
-     * @expectedException YOURLS\Exceptions\ConfigException
-     * @expectedExceptionMessageRegExp /User defined config not found at '[0-9a-z]+'/
 	 */
     public function test_incorrect_config() {
+        $this->expectException(YOURLS\Exceptions\ConfigException::class);
+        $this->expectExceptionMessageMatches('/User defined config not found at \'[0-9a-z]+\'/');
+
         $test = new \YOURLS\Config\Config(rand_str());
         $test->find_config();
     }
 
 	/**
 	 * Test config not found
-     * @expectedException YOURLS\Exceptions\ConfigException
-     * @expectedExceptionMessage Cannot find config.php. Please read the readme.html to learn how to install YOURLS
 	 */
     public function test_not_found_config() {
+        $this->expectException(YOURLS\Exceptions\ConfigException::class);
+        $this->expectExceptionMessage('Cannot find config.php. Please read the readme.html to learn how to install YOURLS');
+
         $test = new \YOURLS\Config\Config();
         $test->set_root(rand_str());
         $test->find_config();
