@@ -5,21 +5,21 @@
  *
  * @group install
  */
-class Install_Version_Tests extends PHPUnit_Framework_TestCase {
+class Install_Version_Tests extends PHPUnit\Framework\TestCase {
 
     protected $ydb_copy;
 
     /**
      * Make a copy of $ydb
      */
-    public function setUp() {
+    public function setUp(): void {
         $this->ydb_copy = yourls_get_db();
     }
 
     /**
      * Restore original $ydb
      */
-    public function tearDown() {
+    public function tearDown(): void {
         yourls_set_db($this->ydb_copy);
     }
 
