@@ -6,9 +6,9 @@
  * @group l10n
  * @since 0.1
  */
-class Translation_Domain_Tests extends PHPUnit_Framework_TestCase {
+class Translation_Domain_Tests extends PHPUnit\Framework\TestCase {
 
-    public static function tearDownAfterClass() {
+    public static function tearDownAfterClass(): void {
         yourls_unload_textdomain( 'test' );
         yourls_unload_textdomain( 'default' );
     }
@@ -26,11 +26,12 @@ class Translation_Domain_Tests extends PHPUnit_Framework_TestCase {
     /**
      * Load custom fake domains - should raise an error
      *
-     * @expectedException PHPUnit_Framework_Error
-     * @expectedExceptionMessageRegExp /Cannot read file [0-9a-z]+\/[0-9a-z]+-fr_FR\.mo\. Make sure there is a language file installed. More info: http:\/\/yourls\.org\/translations/
      * @since 0.1
      */
     public function test_custom_fake_domain() {
+        $this->expectError();
+        $this->expectErrorMessageMatches('/Cannot read file [0-9a-z]+\/[0-9a-z]+-fr_FR\.mo\. Make sure there is a language file installed. More info: http:\/\/yourls\.org\/translations/');
+
         yourls_load_custom_textdomain( rand_str(), rand_str() );
     }
 
