@@ -7,9 +7,9 @@
  * @group idn
  */
 
-class YOURLS_Link_Tests extends PHPUnit_Framework_TestCase {
+class YOURLS_Link_Tests extends PHPUnit\Framework\TestCase {
 
-    protected function tearDown() {
+    protected function tearDown(): void {
         yourls_remove_all_filters( 'get_yourls_site' );
     }
 
@@ -17,7 +17,7 @@ class YOURLS_Link_Tests extends PHPUnit_Framework_TestCase {
      * Check yourls_get_yourls_site() returns a string
      */
     public function test_yourls_site() {
-        $this->assertInternalType("string", yourls_get_yourls_site());
+        $this->assertIsString(yourls_get_yourls_site());
 
         $scheme = yourls_get_protocol( yourls_get_yourls_site() );
         $this->assertContains( $scheme, array( 'http://', 'https://' ), "yourls_get_yourls_site() isn't http(s)://" );
