@@ -5,11 +5,11 @@
  *
  * @group stats
  */
- 
-class Misc_Stats_Tests extends PHPUnit_Framework_TestCase {
+
+class Misc_Stats_Tests extends PHPUnit\Framework\TestCase {
 
     public function test_do_log_redirect() {
-        $this->assertInternalType("bool", yourls_do_log_redirect());        
+        $this->assertIsBool(yourls_do_log_redirect());
     }
 
 }
