@@ -13,8 +13,9 @@
 
 namespace YOURLS\Database;
 
-use YOURLS\Admin\Logger;
-use Aura\Sql\ExtendedPdo;
+use \Aura\Sql\ExtendedPdo;
+use \YOURLS\Database\Profiler;
+use \YOURLS\Database\Logger;
 use PDO;
 
 class YDB extends ExtendedPdo {
@@ -172,12 +173,21 @@ class YDB extends ExtendedPdo {
      * Start a Message Logger
      *
      * @since  1.7.3
-     * @see    \YOURLS\Admin\Logger
-     * @see    \Aura\Sql\Profiler
+     * @see    \Aura\Sql\Profiler\Profiler
+     * @see    \Aura\Sql\Profiler\MemoryLogger
      * @return void
      */
     public function start_profiler() {
-        $this->profiler = new Logger($this);
+        // Instantiate a custom logger and make it the profiler
+        $yourls_logger = new Logger();
+        $profiler = new Profiler($yourls_logger);
+        $this->setProfiler($profiler);
+
+        /* By default, make "query" the log level. This way, each internal logging triggered
+         * by Aura SQL will be a "query", and logging triggered by yourls_debug() will be
+         * a "message". See includes/functions-debug.php:yourls_debug()
+         */
+        $profiler->setLoglevel('query');
     }
 
     /**
@@ -347,42 +357,14 @@ class YDB extends ExtendedPdo {
     /**
      * Return SQL queries performed
      *
-     * Aura\Sql\Profiler logs every PDO command issued. But depending on PDO::ATTR_EMULATE_PREPARES, some are
-     * actually sent to the mysql server or not :
-     *  - if PDO::ATTR_EMULATE_PREPARES is true, prepare() statements are not sent to the server and are performed
-     *    internally, so they are removed from the logger
-     *  - if PDO::ATTR_EMULATE_PREPARES is false, prepare() statements are actually performed by the mysql server,
-     *    and count as an actual query
-     *
-     * Resulting array is something like:
-     *   array (
-     *      0 => array (
-     *           'duration' => 1.0010569095611572265625,
-     *           'function' => 'connect',
-     *           'statement' => NULL,
-     *           'bind_values' => array (),
-     *           'trace' => ...back trace...,
-     *       ),
-     *       // key index might not be sequential if 'prepare' function are filtered out
-     *       2 => array (
-     *           'duration' => 0.000999927520751953125,
-     *           'function' => 'perform',
-     *           'statement' => 'SELECT option_value FROM yourls_options WHERE option_name = :option_name LIMIT 1',
-     *           'bind_values' => array ( 'option_name' => 'test_option' ),
-     *           'trace' => ...back trace...,
-     *       ),
-     *   );
-     *
      * @since  1.7.3
      * @return array
      */
     public function get_queries() {
-        $queries = $this->getProfiler()->getProfiles();
+        $queries = $this->getProfiler()->getLogger()->getMessages();
 
-        if ($this->get_emulate_state()) {
-            // keep queries if $query['function'] != 'prepare'
-            $queries = array_filter($queries, function($query) {return $query['function'] !== 'prepare';});
-        }
+        // Only keep messages that start with "SQL "
+        $queries = array_filter($queries, function($query) {return substr( $query, 0, 4 ) === "SQL ";});
 
         return $queries;
     }
@@ -476,7 +458,7 @@ class YDB extends ExtendedPdo {
         return $this->fetchValue($query);
     }
 
-    public function query($query) {
+    public function query($query, ...$unused) {
         yourls_deprecated_function( '$ydb->'.__FUNCTION__, '1.7.3', 'PDO' );
         yourls_debug_log('LEGACY SQL: '.$query);
         return $this->fetchAffected($query);
