@@ -11,15 +11,15 @@ require_once( 'login_base.php' );
  */
 class Auth_Login_API_Secure_Tests extends Login_Base {
 
-    public static function setUpBeforeClass() {
+    public static function setUpBeforeClass(): void {
         yourls_add_filter( 'is_API', 'yourls_return_true' );
-        $_REQUEST['signature'] = yourls_auth_signature( 'yourls' ); 
+        $_REQUEST['signature'] = yourls_auth_signature( 'yourls' );
         /* Attempt login with valid signature & timestamp. Tests with invalid signatures are made directly
          *  against the check function, not in a full login procedure. See auth.php
          */
     }
 
-    public static function tearDownAfterClass() {
+    public static function tearDownAfterClass(): void {
         yourls_remove_filter( 'is_API', 'yourls_return_true' );
         unset( $_REQUEST['signature'] );
     }
